requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
cffi==1.15.1
charset-normalizer==2.1.1
discord.py==2.0.1
frozenlist==1.3.1
idna==3.4
multidict==6.0.2
numpy==1.23.3
pandas==1.5.0
pycparser==2.21
PyNaCl==1.5.0
python-dateutil==2.8.2
pytz==2022.4
six==1.16.0
yarl==1.8.1
